pyproject.toml
[project]
name = "mise"
version = "0.1.0"
description = "MISE — Live Kitchen Intelligence Agent for the Gemini Live Agent Challenge"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.5.0",
    "google-genai>=1.14.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "websockets>=15.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-secret-manager>=2.21.0",
    "google-cloud-logging>=3.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
